LPS-54620 moves java
