+++ modules/apps/portlet-configuration-icons/portlet-configuration-icons-web/src/com/liferay/portlet/configuration/icons/web/ConfigurationPortletConfigurationIcon.java
@@ -12,10 +12,12 @@
  * details.
  */
 
-package com.liferay.portal.kernel.portlet.configuration;
+package com.liferay.portlet.configuration.icons.web;
 
+import com.liferay.portal.kernel.language.LanguageUtil;
 import com.liferay.portal.kernel.portlet.configuration.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIcon;
+import com.liferay.portal.kernel.util.HtmlUtil;
 import com.liferay.portal.theme.PortletDisplay;
 
 import org.osgi.service.component.annotations.Component;
@@ -26,56 +28,62 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true, service = PortletConfigurationIcon.class
 )
-public class MinimizePortletConfigurationIcon
+public class ConfigurationPortletConfigurationIcon
 	extends BasePortletConfigurationIcon {
 
 	@Override
 	public String getCssClass() {
-		return "portlet-minimize portlet-minimize-icon";
+		return "portlet-configuration portlet-configuration-icon";
 	}
 
 	@Override
 	public String getImage() {
-		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
-
-		if (portletDisplay.isStateMin()) {
-			return "../aui/resize-vertical";
-		}
-
-		return "../aui/minus";
+		return "../aui/wrench";
 	}
 
 	@Override
 	public String getMessage() {
-		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
+		return "configuration";
+	}
 
-		if (portletDisplay.isStateMin()) {
-			return "restore";
-		}
-
-		return "minimize";
+	@Override
+	public String getMethod() {
+		return "get";
 	}
 
 	@Override
 	public String getOnClick() {
 		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
 
-		return "Liferay.Portlet.minimize('#p_p_id_".concat(
-			portletDisplay.getId()).concat("_', this); return false;");
+		StringBuilder sb = new StringBuilder(11);
+
+		sb.append("Liferay.Portlet.openWindow('#p_p_id_");
+		sb.append(portletDisplay.getId());
+		sb.append("_', '");
+		sb.append(portletDisplay.getId());
+		sb.append("', '");
+		sb.append(HtmlUtil.escapeJS(portletDisplay.getURLConfiguration()));
+		sb.append("', '");
+		sb.append(portletDisplay.getNamespace());
+		sb.append("', '");
+		sb.append(LanguageUtil.get(_themeDisplay.getLocale(), "configuration"));
+		sb.append("'); return false;");
+
+		return sb.toString();
 	}
 
 	@Override
 	public String getURL() {
 		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
 
-		return portletDisplay.getURLMin();
+		return portletDisplay.getURLConfiguration();
 	}
 
 	@Override
 	public boolean isShow() {
 		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
 
-		return portletDisplay.isShowMinIcon();
+		return portletDisplay.isShowConfigurationIcon();
 	}
 
 	@Override
