+++ modules/apps/portlet-configuration-icons/portlet-configuration-icons-web/src/com/liferay/portlet/configuration/icons/web/RefreshPortletConfigurationIcon.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.portlet.configuration;
+package com.liferay.portlet.configuration.icons.web;
 
 import com.liferay.portal.kernel.portlet.configuration.BasePortletConfigurationIcon;
 import com.liferay.portal.kernel.portlet.configuration.PortletConfigurationIcon;
@@ -26,35 +26,44 @@ import org.osgi.service.component.annotations.Component;
 @Component(
 	immediate = true, service = PortletConfigurationIcon.class
 )
-public class HelpPortletConfigurationIcon extends BasePortletConfigurationIcon {
+public class RefreshPortletConfigurationIcon
+	extends BasePortletConfigurationIcon {
 
 	@Override
 	public String getCssClass() {
-		return "portlet-help portlet-help-icon";
+		return "portlet-refresh portlet-refresh-icon";
 	}
 
 	@Override
 	public String getImage() {
-		return "../aui/question-sign";
+		return "../aui/refresh";
 	}
 
 	@Override
 	public String getMessage() {
-		return "help";
+		return "refresh";
+	}
+
+	@Override
+	public String getOnClick() {
+		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
+
+		return "Liferay.Portlet.refresh('#p_p_id_".concat(
+			portletDisplay.getId()).concat("_'); return false;");
 	}
 
 	@Override
 	public String getURL() {
 		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
 
-		return portletDisplay.getURLHelp();
+		return portletDisplay.getURLRefresh();
 	}
 
 	@Override
 	public boolean isShow() {
 		PortletDisplay portletDisplay = _themeDisplay.getPortletDisplay();
 
-		return portletDisplay.isShowHelpIcon();
+		return portletDisplay.isShowRefreshIcon();
 	}
 
 	@Override
